Towards #238: First geolocation guesses work
Doing the guessing in the UI, so that I can later redo it at
some intervals.

Still a problem that my home computer gets a very different location
from my phone, so if the last entry is from my phone, the computer
finds Fermentoren to be closer. Just using the last entry seems not
to be enough.