Fix geoloc default
Remember the loc from the last entry, and put it back into
geolocation if nothing closer found. That way, if I walk away
from a bar, at some point the geoloc reverts to showing the
latest entry, and not guessing the new closest bar.

Also, reduced the acceptable distance to 200m, my phone seems to
be well precise enough for that.