FIx #277: Better input form
Now allows recording a new beer, or editing the last beer (or the
one selected via an 'edit' link).

Had to move input elements around a bit to make room for the new
button, and to make the rating and menu selects narrower.

There may still be some corner cases, but it seems to work for now