Detect dev version, adjust display
If the script lives in a directory like beertracker-dev, or is not called
index.cgi, it is assumed to be a development version. In that case we get
a blue background, and a different page title and icon, and a small section
on the top of the page that tells the modification time of the script. If
the production version is newer, prints a warning, you probably should do
a git pull.