Server-side field-hiding logic
Now refreshes the page when changing the type, doing all logic on
server side. Added a Night type line to test. Seems to work,
even though it still produces a pile of warnings when reading
such a line, trying to normalize non-existing fields, and when
displaying it in the long list, probably on every other page
as well.