Display S/L for Fermentoren
The back end will guess the sizes, but we should not display
them when we don't know. Will have to do on the expanded button
on the board, to make sense of drinks.

Moved the sorting of sizes into the scraper script for øb, can
not do numerical sort for S/L