Towards 423: Yearly stats
For some reason the numbers are not exactly the same as in the old
version. But they look right, and I don't need to be so precise
anyway... Added also a count of visits for the year, which can
well be over 365, since I can visit multiple places in a day.