Towards #442: Sorting works.
Removed a lot of debug output, which slowed things down a lot

The table head looks messy. Added placeholders for filter inputs,
that is better, but now it repeats the headers.

Need to sort out a way to add sorting in the filter inputs. Maybe
+ and - for asc/desc sorting? And to make sure we have inputs for
all that we may want to sort/filter by