Fix #456 Ratings stats
Heavy use of chatgtp, made a new module in a couple of hours.
Might have been almost as fast doing most of it manually.

Still needs a bit of tuning for the layout

Making two graphs, one with gnuplot the way I used to do it,
one with graph.js.