Fix #480: Use 'localtime' with 'now'
Time stamps are already in the users local time, but when
asking for the DB for the date, we need to take its idea of
local time, subtract the usual 06:00 to change date late in
the night.

This will fail if I travel to a very different time zone, or
get users in such zones.